Hardcode checksum
Since the `default.conf` file in the unprivileged images will never be the same as the one on the package repositories, the checksums in the entrypoint script will never match and IPv6 will never be enabled by default. By hardcoding the test checksum to the checksum of the unprivileged modified `default.conf` the checksums in the entrypoint script will match when appropriate and the script will run as intended again.